added endpint of getting all doctors of hospital

+++ backend/src/main/java/com/personal/patient/account/models/CreatingDoctorResponse.java
@@ -20,7 +20,7 @@ public class CreatingDoctorResponse {
 
     private String gender;
 
-    private Integer hospitalId;
+    private Long hospitalId;
 
     private List<CreatingSpecialization> specializations;
